requirements.txt
Flask>=2.2.2
gunicorn==20.0.4
requests==2.24.0
python-dotenv==0.21.1